Do not fetch ContentViewhelper nodes recursively
Content ViewHelpers should only be fetched as direct children,
not recursively. Fetching them recursively in templates will
not work if components are nested, as slots of the inner component
would also be applied to the outer component. Placing content in an
if-else block, thus defining the content twice, will also not work
and would no longer be possible with this change, which should
prevent some headaches.